Refactor: combined common/ into one script
