--- app_v2/common/colors.py
@@ -1,9 +1,0 @@
-class Colors:
-    GREY = "#3f3f3f"
-    YELLOW = "#FEDD00"
-    PURPLE = "#6A0DAD"
-    ORANGE = "#F3742B"
-    DARK_ORANGE = "#B83A14"
-    BLUE_PURPLE = "#231650"
-    PINK = "#B52F9C"
-    BLUE = "#1D1BF9"
